refactor base table
