+++ src/components/Table/BaseTable.js
@@ -3,12 +3,9 @@ import { useTranslation } from "react-i18next"
 import { DataGrid } from "@mui/x-data-grid"
 import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFns"
 import { LocalizationProvider } from "@mui/x-date-pickers"
-import { dataGridStyles, TableBox } from "./styled"
-import { useEffect, useMemo, useState } from "react"
-import { USER_STATUS } from "../../utils/constants"
-import TableToolBar from "./TableToolBar"
-import { LinearProgress } from "@mui/material"
-import { Box } from "@mui/system"
+import { TableBox } from "./styled"
+import { useState } from "react"
+import useBaseTableConfig from "./useBaseTableConfig"
 
 const BaseTable = (props) => {
   const { tableConfig } = props
@@ -21,33 +18,14 @@ const BaseTable = (props) => {
     handleSortChange,
     onRowModesModelChange,
     rowModesModel,
-    //tableSource,
     rows,
     setRows,
     ...otherTableConfig
   } = tableConfig
+
   const { t } = useTranslation()
   const [searchInput, setSearchInput] = useState("")
-
-  const localizedTextMap = useMemo(
-    () => ({
-      toolbarFilters: t("toolbarFilters"),
-      filterPanelColumns: t("filterPanelColumns"),
-      filterPanelOperators: t("filterPanelOperators"),
-      filterPanelInputLabel: t("filterPanelInputLabel"),
-      filterOperatorContains: t("filterOperatorContains"),
-      toolbarColumns: t("toolbarColumns"),
-      columnsPanelHideAllButton: t("columnsPanelHideAllButton"),
-      columnsPanelShowAllButton: t("columnsPanelShowAllButton"),
-      columnsPanelTextFieldLabel: t("columnsPanelTextFieldLabel"),
-      columnsPanelTextFieldPlaceholder: t("columnsPanelTextFieldPlaceholder"),
-      filterOperatorIs: t("filterOperatorIs"),
-      noRowsLabel: t("noRowsLabel"),
-      errorOverlayDefaultLabel: t("errorOverlayDefaultLabel"),
-      noResultsOverlayLabel: t("noResultsOverlayLabel"),
-    }),
-    [t]
-  )
+  const baseTableConfig = useBaseTableConfig()
 
   const handleRowEditStart = (params, event) => {
     event.defaultMuiPrevented = true
@@ -98,39 +76,16 @@ const BaseTable = (props) => {
     handleResetTableConfig(handleResetTable)
   }
 
-  const baseTableConfig = {
-    pagination: true,
-    paginationMode: "server",
-    sortingMode: "server",
-    filterMode: "server",
-    disableColumnMenu: true,
-    headerHeight: 40,
-    components: {
-      Toolbar: TableToolBar,
-      LoadingOverlay: LinearProgress,
-    },
-    getRowSpacing: (params) => ({
-      top: params.isFirstVisible ? 10 : 5,
-      bottom: params.isLastVisible ? 0 : 5,
-    }),
-    disableSelectionOnClick: true,
-    editMode: "row",
-    //checkboxSelection
-  }
-
   return (
     <>
       <TableBox>
         <LocalizationProvider dateAdapter={AdapterDateFns}>
           <DataGrid
-            experimentalFeatures={{ newEditingApi: true }}
             {...otherTableConfig}
             {...baseTableConfig}
             rows={rows}
             rowModesModel={rowModesModel}
-            onRowModesModelChange={(newModeModel) =>
-              onRowModesModelChange(newModeModel)
-            } //
+            onRowModesModelChange={onRowModesModelChange}
             componentsProps={{
               toolbar: {
                 searchInput,
@@ -146,28 +101,13 @@ const BaseTable = (props) => {
               pagination: {
                 labelRowsPerPage: t("pageSize"),
               },
-            }} //
-            //loading={true}
-            sx={dataGridStyles} //
-            //checkboxSelection
-            localeText={localizedTextMap} //
-            processRowUpdate={processRowUpdate}
-            onRowEditStart={handleRowEditStart} //
-            onRowEditStop={handleRowEditStop} //
-            onProcessRowUpdateError={handleProcessRowUpdateError}
-            onFilterModelChange={handleFilterModelChange} //
-            onSortModelChange={handleSortModelChange} //
-            isCellEditable={(params) => {
-              const { row, field, colDef, formattedValue } = params
-              if (
-                (field === "email" && !row.isNew) ||
-                (field === "status" && formattedValue === USER_STATUS.temporary)
-              ) {
-                return false
-              } else {
-                return colDef.editable()
-              }
             }}
+            processRowUpdate={processRowUpdate}
+            onRowEditStart={handleRowEditStart}
+            onRowEditStop={handleRowEditStop}
+            onProcessRowUpdateError={handleProcessRowUpdateError}
+            onFilterModelChange={handleFilterModelChange}
+            onSortModelChange={handleSortModelChange}
           />
         </LocalizationProvider>
       </TableBox>
